refactor: 자동차이름 입력, 검증 기능 CarNamesInput 컴포넌트로 이동
