+++ index.html
@@ -14,7 +14,7 @@
         올바른 예) east,west,south,north <br />
       </p>
       <div class="car-game-container">
-        <div>
+        <div class="car-names-container">
           <input id="car-names-input" type="text" />
           <button id="car-names-submit">확인</button>
         </div>
